+++ uCOS-III_Wrapper/uCOS-III/os.h
@@ -758,6 +758,8 @@ struct  os_mutex {
     CPU_CHAR           *NamePtr;
     OS_OBJ_TYPE         Type;
     OS_NESTING_CTR      OwnerNestingCtr;                   /* Mutex is available when the counter is 0               */
+    OS_TCB             *OwnerTCBPtr;
+    OS_PRIO             OwnerOriginalPrio;
 #if OS_CFG_DBG_EN > 0u
     OS_MUTEX           *DbgPrevPtr;
     OS_MUTEX           *DbgNextPtr;